Completed desafío 2
The running system is the copy of the original proposal. The original proposal is in the file desafio 2.ino. The copy is in the file desafio 2 copy.ino. The running system only checks for maximum and minimum colors, if red is max it turns right, if blue is min it turns left.